Add fragments, fix zooming, Cleanup code & add comments.
